move non-exercise fileupload related functionality from "fileupload" to "core" module
