+++ src/main/java/de/tum/cit/aet/artemis/core/service/FileService.java
@@ -58,11 +58,11 @@ import org.springframework.web.multipart.MultipartFile;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.ibm.icu.text.CharsetDetector;
 
+import de.tum.cit.aet.artemis.core.domain.FilePathInformation;
 import de.tum.cit.aet.artemis.core.exception.BadRequestAlertException;
 import de.tum.cit.aet.artemis.core.exception.FilePathParsingException;
 import de.tum.cit.aet.artemis.core.exception.InternalServerErrorException;
 import de.tum.cit.aet.artemis.core.util.CommonsMultipartFile;
-import de.tum.cit.aet.artemis.fileupload.domain.FilePathInformation;
 
 @Profile(PROFILE_CORE)
 @Service
