+++ src/test/java/de/tum/cit/aet/artemis/fileupload/service/FileUploadScheduleServiceTest.java
@@ -15,8 +15,10 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import de.tum.cit.aet.artemis.fileupload.domain.FileUpload;
-import de.tum.cit.aet.artemis.fileupload.repository.FileUploadRepository;
+import de.tum.cit.aet.artemis.core.domain.FileUpload;
+import de.tum.cit.aet.artemis.core.repository.FileUploadRepository;
+import de.tum.cit.aet.artemis.core.service.file.FileUploadScheduleService;
+import de.tum.cit.aet.artemis.core.service.file.FileUploadService;
 
 @ExtendWith(MockitoExtension.class)
 class FileUploadScheduleServiceTest {
