+++ src/test/java/de/tum/cit/aet/artemis/fileupload/service/FileUploadServiceTest.java
@@ -21,10 +21,11 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
+import de.tum.cit.aet.artemis.core.domain.FileUpload;
+import de.tum.cit.aet.artemis.core.domain.FileUploadEntityType;
+import de.tum.cit.aet.artemis.core.repository.FileUploadRepository;
 import de.tum.cit.aet.artemis.core.service.FileService;
-import de.tum.cit.aet.artemis.fileupload.domain.FileUpload;
-import de.tum.cit.aet.artemis.fileupload.domain.FileUploadEntityType;
-import de.tum.cit.aet.artemis.fileupload.repository.FileUploadRepository;
+import de.tum.cit.aet.artemis.core.service.file.FileUploadService;
 
 @ExtendWith(MockitoExtension.class)
 class FileUploadServiceTest {
